Move favicons into common/app
Close #942
