+++ gulp/favicon-inject.js
@@ -6,5 +6,5 @@ gulp.task('favicon-inject', () => {
   const content = `/* eslint-disable quote-props, quotes */
 export default ${JSON.stringify(injectFavicon(), null, 2)};
 `;
-  fs.writeFileSync('./src/browser/app/favicon.js', content);
+  fs.writeFileSync('./src/common/app/favicon.js', content);
 });
