+++ src/common/app/favicon.js
