+++ src/common/app/favicons/browserconfig.xml
