+++ src/common/app/favicons/favicon.ico
