+++ src/common/app/favicons/manifest.json
