+++ src/common/app/favicons/original/favicon.png
